More flexible docker launch scripts

+++ docker/docker.sh
@@ -3,12 +3,13 @@
 UP1_DIR="$(cd "$(dirname "${BASH_SOURCE[0]}")"/.. >/dev/null && pwd)"
 THIS_DIR="$(cd "$(dirname "${BASH_SOURCE[0]}")" >/dev/null && pwd)"
 IMAGE_NAME=`cd "${UP1_DIR}" && echo ${PWD##*/} | tr '[:upper:]' '[:lower:]' | sed -e 's/[-_ ]//g'`
-IMAGE_NAME="${IMAGE_NAME}-jupyter"
 JUPYTER_PORT=8888
 TENSORBORAD_PORT=6006
+DOCKER_RUN_FLAGS="it"
+FOREVER=0
+BUILD=1
+KILL=1
 CMD=""
-
-echo "Building ${IMAGE_NAME}..."
 
 # process named arguments
 while [ $# -gt 0 ]; do
@@ -18,6 +19,18 @@ while [ $# -gt 0 ]; do
       ;;
     --tensorboard_port=*)
       TENSORBORAD_PORT="${1#*=}"
+      ;;
+    --image_suffix=*)
+      IMAGE_NAME="${IMAGE_NAME}-${1#*=}"
+      ;;
+    --forever)
+      DOCKER_RUN_FLAGS+="d"
+      ;;
+    --no-kill)
+      KILL=0
+      ;;
+    --no-build)
+      BUILD=0
       ;;
     --help)
       echo "Usage: docker.sh [--jupyter_port=####|8888] [--tensorboard_port=####|6006] [--help] [command]"
@@ -29,9 +42,24 @@ while [ $# -gt 0 ]; do
   shift
 done
 
-docker kill "${IMAGE_NAME}"
+if [ $KILL -ge 1 ]
+  then
+    echo "Killing ${IMAGE_NAME}..."
+    docker kill "${IMAGE_NAME}"
+fi
 
-docker build -f "${THIS_DIR}/Dockerfile" -t $IMAGE_NAME "${UP1_DIR}" && \
-docker run --gpus=all --rm ${_OPTIONS_:-'-it'} -p $JUPYTER_PORT:8888  -p $TENSORBORAD_PORT:6006 \
-	-v "${UP1_DIR}:/app" --name="${IMAGE_NAME}" \
-	$IMAGE_NAME $CMD
+if [ $BUILD -ge 1 ]
+  then
+    echo "Building ${IMAGE_NAME}..."
+    docker build -f "${THIS_DIR}/Dockerfile" -t $IMAGE_NAME "${UP1_DIR}" || exit 1
+fi
+
+# only map jupyter/tensorboard ports if command is not specified
+if [ -z "$CMD" ]
+  then
+    PORT_MAPPINGS="-p ${JUPYTER_PORT}:8888  -p ${TENSORBORAD_PORT}:6006"
+fi
+
+docker run --gpus=all --rm "-${DOCKER_RUN_FLAGS}" --name="${IMAGE_NAME}" \
+  -v "${UP1_DIR}:/app" $PORT_MAPPINGS \
+  $IMAGE_NAME $CMD
